AxiosErrorHandlers unified in one file and moved to utils
