--- web/src/components/utils/axios_errors_handlers/AxiosErrorHander.ts
@@ -1,9 +1,0 @@
-import Axios from "axios";
-
-const axiosErrorExceptionHandler = (err: Error) => {
-    if (!Axios.isCancel(err)) {
-        console.log(err);
-      }
-}
-
-export default axiosErrorExceptionHandler;
