--- web/src/components/utils/axios_errors_handlers/AxiosErrorHandlerWithCallback.ts
@@ -1,10 +1,0 @@
-import Axios from "axios";
-
-const axiosErrorHandlerWithCallback = (err: Error, callback: () => void) => {
-    if (!Axios.isCancel(err)) {
-        console.log(err);
-        callback();
-      }
-}
-
-export default axiosErrorHandlerWithCallback;
